Workaround for re-running digitization on old MC
Before July 2022 the Scifi signal speed was not determined, thus missing from the
geofiles. At that time it was not a problem for digitization cause light propagation
in the fibers was not simulated. The latter effect was added in Nov 2023. Ever since,
if one re-digitizes old MC, it is essential to have that parameter set.