G4conf: Remove neutron and hadronic process verbosity settings
/mcPhysics/g4NeutronHPVerbose 0
breaks code execution with  GEANT4 v11.4.beta.
The impact of the two removed verbosity settings on the log using the previous v11.1.0 is none. The original commit, which includes both settings, dates 11 years back. In the meantime, NeutronHP merged to ParticleHP and seems support for pure NeutronHP is (perhaps gradually) being dropped.